add distributor page

+++ database/migrations/2024_06_12_150304_create_orders_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('orders', function (Blueprint $table) {
             $table->id();
             $table->date('date');
-            $table->decimal('total_price', 10, 2);
+            $table->unsignedInteger('total_price');
             $table->foreignId('distributor_id');
             $table->timestamps();
