+++ database/migrations/2024_06_12_150320_create_order_details_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('order_id');
             $table->foreignId('product_id');
             $table->integer('quantity');
-            $table->decimal('price', 10, 2);
+            $table->unsignedInteger('price');
             $table->decimal('total_price', 10, 2);
             $table->timestamps();
